update homeware.h
When using a Thermostat i.e. the leght of readed string is longer than 200 chars because the use of very large descriptions and changing the data type from int to char. When using int you can not store a decimal temperature as 21.5 degrees. Passing the datatype to string you store "21.5" and works fine for me.